Check-http can return whole response when failing
By providing a ```whole_response``` boolean parameter to the ```check-http.rb``` script, the whole response will be appended to the body of the message which will be displayed in sensu dashboard. The default value in the new parameter is false, so if you don't specify it, the script will work just as it used to before the change.