Update check-http.rb to fix connection timeout bug
Net:HTTP has an implicit timeout value of 60 seconds for most things. If you run this script with a --timeout greater than 60 seconds, Net::HTTP will act as if the timeout was still 60 seconds at the connection level.

This patch tells Net::HTTP to adjust its timeout so that it's never less than the script --timeout value.