Get the CI tests working again
I'm not sure who introduced what change, but the test failures were
because Sensu's embedded Ruby environment and RVM in TravisCI were
stomping on each other's environment variables.

After doing a bunch of test runs in Travis, this seems to be enough to
get them going again. And has the added benefit of using RVM's Ruby for
the tests, so they cover all the versions of Ruby being tested against.