metrics-curl: force the locale to format numbers correctly
Some locale (like the French "fr_FR" one) use a comma as a fixed/decimal
number separator.
This breaks the parsing of curl metrics which are also
expected to be separated by a comma.

Forcing the locale to a very sane default such as the C locale forces the
number formatting to use a dot as number separator.